+++ src/localisation/english.ts
@@ -421,6 +421,7 @@ const ENGLISH: Translations = {
   [Phrase.RecordCurrentRaidsOnlyDescription]: "Only record raid encounters from the current tier, this only applies to retail raid encounters.",
   [Phrase.UploadCurrentRaidsOnlyLabel]: "Current Tier Only",
   [Phrase.UploadCurrentRaidsOnlyDescription]: "Only upload raid encounters from the current tier, this only applies to retail raid encounters.",
+  [Phrase.MustNotBeEmpty]: "Must not be empty",
 };
 
 export default ENGLISH;